Dispatch undistort
